requirements.txt
torch>=2.0
pdf2image>=1.16
pyzbar>=0.1.9
pillow>=9.0
opencv-python>=4.5
ultralytics>=8.0.20
PyJWT
